refactor(BridgeGameTest): 테스트 코드 일부 분리 및 오타 수정
- 테스트를 하는 코드를 제외한 메서드는 BridgeGameTestTool로 이동함
- BridgeGameTest에서 Tool을 확장하도록 함
- excepted -> expected로 오타 수정함
